Moved App class to Program.cs
