+++ InventoryManager/ConsoleInteraction/IUserInteraction.cs
@@ -2,47 +2,45 @@
 
 namespace InventoryManager.ConsoleInteraction
 {
-
     /// <summary>
-    /// <see cref="IUserInteraction"/> used to interact with the user
+    /// Used to interact with user
     /// </summary>
     public interface IUserInteraction
     {
         /// <summary>
-        /// <see cref="DisplayOptions"/> display all the available <see cref="Product"/>
+        /// Display all the available <see cref="Product"/>
         /// </summary>
         public void DisplayOptions();
 
         /// <summary>
-        /// <see cref="DisplayEditOptions"/> display the edit options
+        /// Display the edit options
         /// </summary>
         public void DisplayEditOptions();
 
         /// <summary>
-        /// <see cref="DisplayAllProducts(List{Product})"/> display all the <see cref="Product"/>
+        /// Display all the <see cref="Product"/>
         /// </summary>
-        /// <param name="productList">all the available <see cref="Product"/></param>
+        /// <param name="productList">Product List</param>
         public void DisplayAllProducts(List<Product> productList);
 
         /// <summary>
-        /// <see cref="GetNewProductDetail"/> get new product details
+        /// Get product details
         /// </summary>
-        /// <returns>New <see cref="Product"/> with deatils</returns>
+        /// <returns>returns product</returns>
         public Product GetNewProductDetail();
 
         /// <summary>
-        /// <see cref="GetAndValidateStringInput(string)"/> validate the user input string
+        /// Validates input string
         /// </summary>
-        /// <param name="message">message to be printed</param>
-        /// <returns>user input string</returns>
+        /// <param name="message">message</param>
+        /// <returns>returns string</returns>
         public string GetAndValidateStringInput(string message);
 
         /// <summary>
-        /// <see cref="GetAndValidateIntInput(string)"/> validate the user input integer
+        ///  Validates input integer
         /// </summary>
-        /// <param name="message">message to be printed</param>
-        /// <returns>user input integer</returns>
+        /// <param name="message">message</param>
+        /// <returns>returns integer</returns>
         public int GetAndValidateIntInput(string message);
-
     }
 }
