+++ InventoryManager/Manager/IManageInventory.cs
@@ -1,32 +1,28 @@
-﻿using InventoryManager.Model;
-
-namespace InventoryManager.Manager
+﻿namespace InventoryManager.Manager
 {
-
     /// <summary>
-    ///  IInverntoryManager interface used to handle the features
+    ///  Used to handle features
     /// </summary>
     public interface IManageInventory
     {
         /// <summary>
-        /// <see cref="AddNewProduct"/> to add new product <see cref="Product"/>
+        /// Add new Product
         /// </summary>
         public void AddNewProduct();
 
         /// <summary>
-        /// <see cref="DeleteExistingProduct"/> deletes the exisiting product <see cref="Product"/>
+        /// Deletes exisiting Product
         /// </summary>
         public void DeleteExistingProduct();
 
         /// <summary>
-        /// <see cref="EditExistingProduct"/> edits the exisiting product <see cref="Product"/>
+        /// Edits  exisiting Product
         /// </summary>
         void EditExistingProduct();
 
         /// <summary>
-        /// <see cref="SearchProduct"/> which searches the product <see cref="Product"/>
+        /// Searches Product
         /// </summary>
         void SearchProduct();
-
     }
 }
